Add D.C. and Fine
The volta bracket for “1” is a bit longer than in the original, but I
haven’t found a way to make it shorter yet. The voltaSpannerDuration
override suggested in the manual doesn’t work and makes LilyPond spew
out a bunch of warnings about “NaN found”, presumably because it’s
trying to apply the duration to the already finished “Fine” bracket
instead of the “1” bracket.