Add tie around Da Capo
The tie at the end of the score is entered as a Laissez Vibrer, since
LilyPond doesn’t have anything else to mirror a \repeatTie. The ties and
the parentheses aren’t perfect yet, but I’ll try to solve that later
once I have Internet access.

The ambitus engraver is moved from the Voice to the Staff context so
that we don’t get a second ambitus for the auxiliary voice used for the
repeated note.