Fix extenders
To make the extenders span the full length of the note, we add a hidden
grace note near the end of the note’s duration – a technique suggested
by Graham King on the lilypond-user mailing list [1]. This is
encapsulated in a new music function, \extendedNote, which applies the
technique to the following note (its argument).

A few extenders were also just missing in the lyrics – apparently I’d
forgotten to enter them.

[1]: https://lists.gnu.org/archive/html/lilypond-user/2018-10/msg00252.html