Fix ambitus
The extenders fix distorted the ambitus, since the invisible grace note
is still counted into the ambitus and, thanks to it being absolute,
could lie outside the rest of the voice. Fix that by using a drum note
for the invisible grace note: drum notes (or any other unpitched note
heads) don’t affect the ambitus.